Refactor scrape code into two modules
