+++ package.json
@@ -9,7 +9,7 @@
     "react-scripts": "2.1.1"
   },
   "scripts": {
-    "scrapeOSMData": "node -r esm ./scrapeOSMData.js",
+    "scrapeOSMData": "node -r esm ./scrapeOSMData/index.js",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
